OTA update refactoring
- Patches renamed into hotfixes.
- add gptokeyb2 to control Terminal with D-PAD
- add shellect.sh (credits: Onion)
- update_common.sh for useful functions (connectivity check, disk space, download, fsck...)
- ota_bootstrap.sh now upgrade the different update scripts
- add update_ota_release.sh, in charge of release download & integrity checks
- add logs
- crossmix_update.sh: better display, use update_common.sh
- App OTA -> launch.sh -> Terminal app + ota_update.sh -> ota_bootstrap.sh, update_ota_release.sh, hotfixes/CrossMix-OS_vx.x.x.sh

+++ _assets/scripts/ota_bootstrap.sh
@@ -1,20 +1,21 @@
 #!/bin/sh
-BOOTSTRAP_VERSION=1.0
+BOOTSTRAP_VERSION=1.1
 GITHUB_REPOSITORY=cizia64/CrossMix-OS
-version=$(cat /mnt/SDCARD/System/usr/trimui/crossmix-version.txt)
+CommonUpdateScript="/mnt/SDCARD/System/usr/trimui/scripts/update_common.sh"
 
-url="https://raw.githubusercontent.com/$GITHUB_REPOSITORY/main/_assets/patchs/CrossMix-OS_v$version.sh"
+if [ ! -f CommonUpdateScript ]; then
+    wget --no-check-certificate --quiet --show-progress -O "$CommonUpdateScript" "https://raw.githubusercontent.com/cizia64/CrossMix-OS/refs/heads/main/System/usr/trimui/scripts/update_common.sh"
+    sync
+    source "$CommonUpdateScript"
 
-if /mnt/SDCARD/System/bin/wget -q --spider "$url"; then
-
-    button=$(/mnt/SDCARD/System/usr/trimui/scripts/infoscreen.sh -m "Patch for CrossMix-OS_v$version found. Press A to apply, B to cancel." -k "A B")
-    if [ "$button" = "A" ]; then
-        curl -k -s "$url" | sh
-    fi
-
-else
-
-    /mnt/SDCARD/System/usr/trimui/scripts/infoscreen.sh -m "No new version, come back later ;)" -t 1
+    download_file "crossmix_update.sh" "https://raw.githubusercontent.com/$GITHUB_REPOSITORY/refs/heads/main/System/usr/trimui/scripts/crossmix_update.sh" -d "/mnt/SDCARD/System/usr/trimui/scripts/"
+    download_file "update_ota_release.sh" "https://raw.githubusercontent.com/$GITHUB_REPOSITORY/refs/heads/main/System/usr/trimui/scripts/update_ota_release.sh" -d "/mnt/SDCARD/System/usr/trimui/scripts/"
+    download_file "ota_update.sh" "https://raw.githubusercontent.com/$GITHUB_REPOSITORY/refs/heads/main/System/usr/trimui/scripts/ota_update.sh" -d "/mnt/SDCARD/System/usr/trimui/scripts/"
+    download_file "launch.sh" "https://raw.githubusercontent.com/$GITHUB_REPOSITORY/refs/heads/main/Apps/OTA-update/launch.sh" -f "/mnt/SDCARD/Apps/OTA-update/launch.sh"
+    download_file "keys.gptk" "https://raw.githubusercontent.com/$GITHUB_REPOSITORY/refs/heads/main/Apps/OTA-update/launch.sh" -f "/mnt/SDCARD/Apps/OTA-update/keys.gptk"
+    download_file "update_common.sh" "https://raw.githubusercontent.com/$GITHUB_REPOSITORY/refs/heads/main/System/usr/trimui/scripts/update_common.sh" -d "/mnt/SDCARD/System/usr/trimui/scripts/"
+    sync
+    /mnt/SDCARD/System/usr/trimui/scripts/infoscreen.sh -m "Update scripts upgraded, please launch OTA update again." -t 5
+    killall -2 SimpleTerminal
 
 fi
-killall -2 SimpleTerminal
